Use consistent integer types.

+++ lib/isce/core/Kernels.h
@@ -110,14 +110,14 @@ class isce::core::NFFTKernel : public isce::core::Kernel<T> {
          * @param[in] n         Length of input signal.
          * @param[in] fft_size  FFT Transform size (> n).
          */
-        NFFTKernel(size_t m, size_t n, size_t fft_size);
+        NFFTKernel(int m, int n, int fft_size);
 
         T operator()(double x) const override;
 
     private:
-        size_t _m;
-        size_t _n;
-        size_t _fft_size;
+        int _m;
+        int _n;
+        int _fft_size;
         T _scale;
         T _b;
 };
@@ -133,13 +133,13 @@ class isce::core::TabulatedKernel : public isce::core::Kernel<T> {
          * @param[in] kernel    Kernel to sample.
          * @param[in] n         Table size.
          */
-        TabulatedKernel(const isce::core::Kernel<T> &kernel, size_t n);
+        TabulatedKernel(const isce::core::Kernel<T> &kernel, int n);
 
         T operator()(double x) const override;
     
     private:
         std::valarray<T> _table;
-        size_t _imax;
+        int _imax;
         T _1_dx;
 };
 
@@ -154,7 +154,7 @@ class isce::core::ChebyKernel : public isce::core::Kernel<T> {
          * @param[in] kernel    Kernel to fit (assumed even).
          * @param[in] n         Number of coefficients.
          */
-        ChebyKernel(const isce::core::Kernel<T> &kernel, size_t n);
+        ChebyKernel(const isce::core::Kernel<T> &kernel, int n);
 
         T operator()(double x) const override;
